refactor: Separate tests for two schema types (#33)
